pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "cjambo",email = "chenjambo@gmail.com"},
]
dependencies = [
    "discord-py>=2.3.1",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

